init some hero aspects
