+++ libs/hmm1-adventure-ui/src/TroopDetailsWindow/TroopDetailsWindow.stories.tsx
@@ -1,13 +1,13 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { creatureById, CreatureId, Luck, Morale, Skill } from '@heroesjs/hmm1-core';
+import { creatureDataById, CreatureId, Luck, Morale, Skill } from '@heroesjs/hmm1-core';
 
 import { TroopDetailsWindow } from './TroopDetailsWindow';
 
 const meta: Meta<typeof TroopDetailsWindow> = {
   args: {
     count: 41,
-    creature: creatureById[CreatureId.Peasant],
+    creature: creatureDataById[CreatureId.Peasant],
     luck: Luck.Normal,
     morale: Morale.Good,
     skillsBonus: {
@@ -27,7 +27,7 @@ export const Primary: Story = {};
 export const Shots: Story = {
   args: {
     count: 10,
-    creature: creatureById[CreatureId.Centaur],
+    creature: creatureDataById[CreatureId.Centaur],
     luck: Luck.Normal,
     morale: Morale.Good,
     skillsBonus: {
