+++ libs/hmm1-core/src/lib/data/index.ts
@@ -1,0 +1,2 @@
+export { type CreatureData, creatureData, creatureDataById } from './creatures';
+export { heroClassData, type HeroClassData, heroClassDataById } from './heroClasses';
